Renaming gm-data files reference.conf files to be loaded by default, created shaded file with all reference.conf files together
