+++ gm-data-gdelt/src/main/resources/reference.conf
@@ -1,7 +1,6 @@
-geomesa.converters = {
-  gdelt = {
-    sft = {
-      type-name = gdelt
+geomesa {
+  sfts {
+    gdelt = {
       fields = [
         { name = globalEventId, type = String }
         { name = eventCode,     type = String }
@@ -38,8 +37,9 @@ geomesa.converters = {
         geomesa.table.sharing = "false"
       }
     }
-
-    converter = {
+  }
+  converters = {
+    gdelt = {
       type = delimited-text
       format = TDF
       id-field = "md5(string2bytes($0))"
