+++ gm-data-geolife/src/main/resources/reference.conf
@@ -1,7 +1,6 @@
-geomesa.converters = {
-  geolife = {
-    sft = {
-      type-name = geolife
+geomesa {
+  sfts {
+    geolife = {
       fields = [
         { name = userId,   type = String }
         { name = altitude, type = Int }
@@ -12,7 +11,9 @@ geomesa.converters = {
         geomesa.table.sharing = "false"
       }
     }
-    converter = {
+  } 
+  converters {
+    geolife = {
       type = delimited-text
       format = DEFAULT
       id-field = "md5(string2bytes($0))"
