+++ gm-data-gtd/src/main/resources/reference.conf
@@ -1,10 +1,9 @@
 // Global Terrorism Database http://www.start.umd.edu/gtd
 // This conf and the ingest are tolerant to missing values in the file, but ingest performance is slower. See README.md for more info.
 // this conf converts date from integer year month and day fields
-geomesa.converters = {
-  gtd = {
-    sft = {
-      type-name = "gtd"
+geomesa {
+  sfts {
+    gtd = {
       attributes = [
         { name = "event_id"       , type = "String" , index = false             }
         { name = "dtg"            , type = "Date"   , index = true              }
@@ -18,8 +17,9 @@ geomesa.converters = {
         { name = "propertydamage" , type = "Double" , index = false             }
       ]
     }
-
-    converter = {
+  } 
+  converters {
+    gtd = {
       type = "delimited-text",
       format = "CSV",
       id-field = "$event_id",
