+++ gm-data-osm-gpx/src/main/resources/reference.conf
@@ -1,7 +1,6 @@
-geomesa.converters = {
-  osm-gpx = {
-    sft = {
-      type-name = osm
+geomesa {
+  sfts {
+    osm-gpx = {
       fields = [
         { name = trackId,   type = String }
         { name = elevation, type = Double }
@@ -12,7 +11,9 @@ geomesa.converters = {
         geomesa.table.sharing = "false"
       }
     }
-    converter = {
+  } 
+  converters {
+    osm-gpx = {
       type = xml
       id-field = "md5(string2bytes(concat(regexReplace('.*?(\\d+)\\.gpx'::r, '$1', $inputFilePath), xml2string($0))))"
       feature-path = "trk/trkseg/trkpt"
@@ -26,5 +27,5 @@ geomesa.converters = {
         { name = geom,                        transform = "point($lon, $lat)" }
       ]
     }
-  }
+  } 
 }
