+++ gm-data-tdrive/pom.xml
@@ -11,5 +11,4 @@
 
     <artifactId>gm-data-tdrive</artifactId>
 
-
 </project>
