+++ gm-data-tdrive/src/main/resources/reference.conf
@@ -1,7 +1,6 @@
-geomesa.converters = {
-  tdrive = {
-    sft = {
-      type-name = tdrive
+geomesa {
+  sfts {
+    tdrive = {
       fields = [
         { name = taxiId, type = String }
         { name = dtg,    type = Date }
@@ -12,7 +11,9 @@ geomesa.converters = {
         table.indexes.enabled = "records,z3,attr_idx"
       }
     }
-    converter = {
+  }  
+  converters {
+    tdrive = {
       type = delimited-text
       format = DEFAULT
       id-field = uuid()
@@ -22,5 +23,5 @@ geomesa.converters = {
         { name = geom,   transform = "point($3::double, $4::double)" }
       ]
     }
-  }
+  } 
 }
