+++ gm-data-twitter/pom.xml
@@ -11,5 +11,4 @@
 
     <artifactId>gm-data-twitter</artifactId>
 
-
 </project>
