+++ gm-data-twitter/src/main/resources/reference.conf
@@ -1,7 +1,6 @@
-geomesa.converters = {
-  twitter = {
-    sft = {
-      type-name = twitter
+geomesa  {
+  sfts {
+    twitter = {
       fields = [
         { name = user_id      , type = String, index = true }
         { name = user_name    , type = String }
@@ -25,7 +24,10 @@ geomesa.converters = {
         // table.indexes.enabled = "records,z3,attr_idx"
       }
     }
-    converter = {
+  }
+  converters {
+    twitter {
+      name = twitter
       type = json
       id-field = "$tweet_id"
       fields = [
